Use MUnit in ParserBaseSuite
